x-telegram-bot/requirements.txt
python-telegram-bot>=20.0
browser-cookie3>=0.19.1
httpx>=0.27.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
